backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
asyncpg==0.30.0
greenlet==3.1.1
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
openpyxl==3.1.5
python-multipart==0.0.20
aiosqlite==0.20.0      # local dev only; remove for Supabase/Render
